Correct types via cast for windows; strlen is type size_t in windows, not int.
git-svn-id: file:///home/jj/hercules.svn/trunk@6010 956126f8-22a0-4046-8f4a-272fa8102e63

+++ hconsole.c
@@ -343,7 +343,7 @@ void translate_keystroke( char kbbuf[], int* pkblen )
 
             } // end switch( ch2 )
 
-            *pkblen = strlen( kbbuf );      // inform caller #of chars
+            *pkblen = (int)strlen( kbbuf );      // inform caller #of chars
             break;
 
         } // end case: 0x00, 0xE0
